feat(core,formatter): moved filesize to formatters dir
