+++ packages/components/core/formatters/filesize/index.ts
@@ -1,3 +1,3 @@
 export * from './config';
-export * from './data-size.pipe';
+export * from './formatter';
 export * from './size';
