+++ packages/components/core/utils/public-api.ts
@@ -1,2 +1,1 @@
-export * from './data-size';
 export * from './utils';
